refactor(shared): centralize page scope contract
