+++ packages/shared/src/contracts/index.ts
@@ -25,6 +25,7 @@ export * from "./inbox.js";
 export * from "./inbox-routes.js";
 export * from "./memory-routes.js";
 export * from "./misc-routes.js";
+export * from "./page-scope.js";
 export * from "./permissions.js";
 export * from "./permissions-routes.js";
 export * from "./personal-assistant.js";
